- Fixed "nc_0" error detecting replicas of files   - Fixed in Harvest   - Fixed in Download   - TODO detect correctly files ("nc_0") in FileChooser that are being downloaded
-Added launch web browser option in JSONViewer (in services)

+++ ESGFToolsUI/src/es/unican/meteo/esgf/download/FileDownloadStatus.java
@@ -821,7 +821,9 @@ public class FileDownloadStatus implements Runnable, Download, Serializable {
         setCurrentSize(0);
 
         // remove file or directory
-        file.delete();
+        if (file != null) {
+            file.delete();
+        }
 
         logger.trace("[OUT] reset");
     }
